fix: return PhysicalError from promote helpers and split physical tests (#2088)
`promote_filed_physical_book` / `promote_filed_physical_books` returned a
raw `sqlx::Error` across the module boundary that rule 02 forbids, while
the sibling `PhysicalError` enum already had a transparent `Sqlx` variant.
Both now return `PhysicalError`, and each call site's error type gains a
transparent `#[from]` variant to carry it:

- `InitDbError::Physical`, alongside the eight sibling-module variants it
  already aggregates.
- `MergeError::Physical` and `SyncError::Physical`, the same shape
  `MergeError` already uses for `Snapshot(#[from] serde_json::Error)`.
  Deliberately not a hand-written `From` that collapses into the `Db`
  variant: that would mean synthesizing a fake `sqlx::Error::Protocol` for
  the non-DB cases, which is the debt #2052 is removing. Nothing matches
  either enum exhaustively — `MergeError`'s only two consumers
  (`frontend/src/rpc/books.rs`) use `Err(e @ (…))` patterns with a
  catch-all — so the new variants break no caller.
- The ebook attach chain (`attach_ebook_file` → `try_attach_new_ebook` /
  `sync_changed_one` → `sync_new` / `sync_changed`) moves from
  `sqlx::Error` to the crate-internal `SyncError`, which is what the
  audiobook sibling already used; both bucket writers are called from
  `anyhow`-returning entry points, so nothing above them changes.

The pre-existing `From<SettingsError>` impls are left alone — they belong
to #2052.

Also splits `db/src/physical/tests.rs` (895 lines, over the ~800-line
file-shape cap) into `tests/{mod,copies,fileless,promote,remove,wishlist}.rs`,
mirroring `server/src/backend/opds/tests/`: shared fixtures (`pool`,
`seed_user`, `count`) live in `mod.rs`, each suite matches the production
module it covers. Pure reorganization — 33 tests before, 33 after. (#2089)
